[refactor] move recordActivity to tripcourse
