+++ app/build.gradle
@@ -34,6 +34,7 @@ android {
     buildFeatures{
         viewBinding = true
     }
+    namespace 'com.olympos.tripbook'
 }
 
 dependencies {
