+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    package="com.olympos.tripbook">
+    xmlns:tools="http://schemas.android.com/tools">
 
     <uses-permission android:name="android.permission.INTERNET"/>
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
@@ -27,7 +26,7 @@
             </intent-filter>
         </activity>
         <activity android:name=".src.trip.TripActivity"/>
-        <activity android:name=".src.trip.RecordActivity"/>
+        <activity android:name=".src.tripcourse.RecordActivity"/>
         <activity android:name=".src.tripcourse.TripCourseActivity"/>
         <activity android:name=".src.user.SignInActivity"/>
         <activity android:name=".src.user.MyPageActivity"/>
