+++ app/src/main/java/com/olympos/tripbook/src/tripcourse/TripCourseActivity.kt
@@ -9,7 +9,6 @@ import com.google.gson.Gson
 import com.olympos.tripbook.R
 import com.olympos.tripbook.config.BaseActivity
 import com.olympos.tripbook.databinding.ActivityTripcourseBinding
-import com.olympos.tripbook.src.trip.RecordActivity
 import com.olympos.tripbook.src.trip.model.Trip
 import com.olympos.tripbook.src.tripcourse.model.TripCourse
 import com.olympos.tripbook.src.tripcourse.view.TripCourseRVAdapter
